Optionally fetch trade history from Binance
Fetching trade history from Binance is a huge pain in the ass, but it's
working. The previous method of providing a JSON file through the
`trades_file` of the TOML configuration is still available. If this
configuration variable is not set, or empty, it will then fetch trade
history from Binance instead.

Binance doesn't offer an endpoint to get all trades. This is a partial
implementation that relies on the `trading_pair` config option. Meaning
it will only support trade history against a single target.

A proper implementation using Binance API would require requesting all
available symbols in Binance (thousands), then request your account data
to figure out which coins you have >= 0 on, and then try to get all
history for all combinations of your coin wallet + Binance symbol. This
is incredibly inefficient, and you'll also be dealing with heavy rate
limits while trying to do this.