Add notification rules and routines
This commit is a bit messy, but basically implements the notification
rules, configurable through the TOML file. This makes it so if the trade
doesn't surpass the rules, it won't be sent to Telegram.

Additionally, it keeps the previous trade in memory, and if the change
is under 1% of difference between previous price and the current, it
will not notify again. This prevents getting spammed with notifications
as soon as the trade goes over the notification threshold.

This commit also implements the fetching of trades (either from JSON, or
from Binance), in a thread, by default set to 5 minutes. The fetching of
current pricing with optional notification is a separate routine running
every 45 seconds.